DCS-271: changes to feature and integration tests following PR review
